Integrated the new message system in all levels
Fixed bug on which death reloads won't occur if the death message is not the same id as the ending message.
Fixed bug on which the next message would start at the transparency of the last message if last message was cut off.
Fixed bug that didn't end the final level upon click.

All in all, this appears to be a cleaner, more efficient system. The chains can be generated from one document and be loaded at any time due to triggers, meaning less hard coding of specific triggers.

Limitation is dynamic changing of dialogue doesn't occur as well, but stories can be built around that limitation.