Added switching objects
Added ability to switch objects using 'q' and 'e'. Will have to update to Unity 6 if I intend to use mouse scroll.

Note: Try to make an inbuilt default update system rather than each level needs it's own setup.

Note: Clean up the fps controller. Lots of different systems, so work on what is good, and try to make it more efficient.

Note: Items and Positions needs to be mentioned separately, maybe make it so that one controls the other?