Added body with head
Created body with head
= Fixed HeldItem in player controller not matching up with inventory held item.
= Fixed doors not behaving as held interactions.
= Fixed some animation issues
+ Added IK for holding items.
+ Added a pickUpPosition to hold items to adjust so the hand will move to the correct position (only implemented on cans so far)

Will monitor for more errors. The inventory system is a very messy addition to the code. It may be best to deprecate the helditem system entirely and always query the inventory script for whether something is in the active hand.

There may be a place to seperate interactions and movement controls into seperate components to better set up player control and allow other creatures to use interactions properly with a seperate controller.